+++ processes/hathiTrust.py
@@ -78,7 +78,9 @@ class HathiTrustProcess(CoreProcess):
 
     @staticmethod
     def returnHathiDateFormat(strDate):
-        if 'T' in strDate:
+        if 'T' and '-' in strDate:
+            return '%Y-%m-%dT%H:%M:%S%z'
+        elif 'T' in strDate:
             return '%Y-%m-%dT%H:%M:%S'
         else:
             return '%Y-%m-%d %H:%M:%S %z'