+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - MagicMock for test_createSearch method
 ### Fixed
 - ES8 API only counting up to 10,000 search results
+- Add bulk retry functionality for ES writes/deletes
 
 ## 2022-04-26 -- v0.10.2
 ### Added