+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## unreleased -- v0.10.3
+## 2022-05-19 -- v0.10.3
 ### Added
 - MagicMock for test_createSearch method in test_api_es.py file
 - Getter and setter method for record model 
@@ -13,6 +13,7 @@
 - Add bulk retry functionality for ES writes/deletes
 - Edition field for some editions now filled
 - HathiTrust data import with new date formats
+- Hotfix for setter method of record model
 
 ## 2022-04-26 -- v0.10.2
 ### Added